repr: change numeric constant types to decrease fallible conversions
Eliminate most fallible conversions when dealing with the constants for
the numeric type, like NUMERIC_DATUM_MAX_PRECISION. The approach is to
leave these constants in the narrowest type that will fit them, and
convert to wider types as is convenient.

+++ src/interchange/src/json.rs
@@ -340,7 +340,7 @@ fn build_row_schema_field<F: FnMut() -> String>(
         }
         ScalarType::Numeric { scale } => {
             let (p, s) = match scale {
-                Some(scale) => (NUMERIC_DATUM_MAX_PRECISION, usize::from(*scale)),
+                Some(scale) => (NUMERIC_DATUM_MAX_PRECISION, *scale),
                 None => (NUMERIC_AGG_MAX_PRECISION, NUMERIC_DATUM_MAX_PRECISION),
             };
             json!({
